HRMS/hrms_proj/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.3
djangorestframework==3.15.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
